requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.5
django-redis==6.0.0
djangorestframework==3.16.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
kombu==5.5.4
outcome==1.3.0.post0
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0